requirements.txt
defusedxml==0.4.1
Django==1.10.2
django-iban-field==0.8
oauthlib==2.0.0
pkg-resources==0.0.0
psycopg2==2.6.2
PyJWT==1.4.2
python-social-auth==0.2.21
python3-openid==3.0.10
requests==2.11.1
requests-oauthlib==0.7.0
six==1.10.0
